Main: VC++ 19.2 update 11 build adjustments (lossy floating point conversions). Also, IPRT does support formatting floating point numbers. bugref:8489
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@93310 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/testcase/tstCollector.cpp
@@ -142,7 +142,7 @@ void measurePerformance(pm::CollectorHAL *collector, const char *pszName, int cV
         /* Process RAM usage */
         N_CALLS(cVMs, getProcessMemoryUsage(processes[call], &tmp));
     }
-    RTPrintf("\n%d VMs -- %.2f%% of CPU time\n", cVMs, (RTTimeNanoTS() - start) / 10000000. / times);
+    RTPrintf("\n%d VMs -- %u%% of CPU time\n", cVMs, (unsigned)((double)(RTTimeNanoTS() - start) / 10000000.0 / times));
 
     /* Shut down fake VMs */
     shutdownProcessList(processes);
